requirements.txt
# Copyright (c) 2021 OM SANTOSHKUMAR MASNE.
# All Rights Reserved.
# Licensed under the MIT license.
# See LICENSE file in the project root for license information.

Flask
Flask-SocketIO
pytorch
numpy
nltk
